requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
Django==4.2.16
fire==0.7.0
fonttools==4.54.1
idna==3.10
language-tool-python==2.8.1
lxml==5.3.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
pdf2docx==0.5.8
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
PyMuPDF==1.24.11
pypdfium2==4.30.0
python-docx==1.1.2
requests==2.32.3
sqlparse==0.5.1
termcolor==2.5.0
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
